Fixed: The screen area used to render the image didn't always correspond with the cockpit/HUD mode chosen Fixed: When changing the ship type, the old ship still got rendered until it got destroyed or the gameplay options menu was opened Fixed: When the Fusion gun was set as built-in weapon, the Phoenix was equipped with it although it should not be able to carry Fusion guns Fixed: Object information could get messed up when loading a save game Improved: Depending on the internal object list usage, savegames can take less space particularly with many objects in a level Fixed: Joining a multiplayer game in progress could take very long or fail altogether Changed: Unless playing in nostalgia node, the automap flight controls now behave like the regular in-mine flight controls New: D2X-XL has an observer mode now, allowing you to either freely roam the map or follow other players between having died and respawning. In observer mode, you are initially placed in the automap. You can toggle between players or free roaming mode by pressing the space bar.This goes in a round robin mode: Free roaming -> player 1 -> player 2 -> ... -> free roaming. In team games, you cannot freely roam the map, but only watch your next spawn location. You also can only follow players from your own team.  Changed: when observing a disconnected player, observer mode will automatically switch to the next connected player or to free roaming mode Changed: When the final boss or reactor of a level is destroyed, players in observer mode will be dropped into the mine
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13504 01821946-851b-0410-b7f4-f1477c70e57a

+++ 2d/rle.cpp
@@ -787,7 +787,7 @@ if (l < 0)
 	return -1;
 if ((l > m_info.props.h * m_info.props.rowSize) && !Resize (l))
 	return -1;
-if (Size () < l) {
+if (Size () < (size_t) l) {
 	if (!Resize (l))
 		return 0;
 	}
